refactor(base_type_copilot): extract tests to sibling (#1266)
- base_type_copilot.rs: 932 -> 579 LOC
- new tests_base_type_copilot.rs (351 LOC)
- widen 5 helper fns to pub(super)
